% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_functions.R
\name{aggregate_to_level_i}
\alias{aggregate_to_level_i}
\title{Aggregate composition data according to CORINE Land Cover levels}
\usage{
aggregate_to_level_i(level_3_df, i)
}
\arguments{
\item{level_3_df}{Data frame of compositions, where the composition
variables are named with three digits. Compatible with the output of
\code{get_full_compositions}.}

\item{i}{1 or 2. Defines the output level of aggregation.}
}
\value{
Data frame similar to \code{level_3_df}. The composition columns are
replaced with data aggregated at the desired output level.
}
\description{
\code{aggregate_to_level_i} transforms a data frame where the composition data
is given at the third level of resolution to the second or first level.
}
\examples{
# CORINE data sets must be loaded first
clc_path <- system.file(
  "extdata", package = "corine.land.cover.landuse.extraction"
)
set_corine_land_cover_path(clc_path)
# Extract with auto selection
level_3_compositions <- get_full_compositions(
  example_points_df, buffer_radius_m = 1, input_clc_year = "auto"
)
# Combine compositions
level_2_compositions <- aggregate_to_level_i(level_3_compositions, 2)
level_1_compositions <- aggregate_to_level_i(level_3_compositions, 1)
level_3_compositions ; level_2_compositions ; level_1_compositions
}
