% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_corine_land_cover_years.R
\name{check_clc_year}
\alias{check_clc_year}
\title{Check if the given year is available}
\usage{
check_clc_year(input_clc_year)
}
\arguments{
\item{input_clc_year}{Integer.}
}
\value{
Nothing if the year is available. An error otherwise.
}
\description{
\code{check_clc_year} returns an error is the given year is not available.
Available years must have be calculated with the function
\code{\link[=set_corine_land_cover_path]{set_corine_land_cover_path()}}.
}
\examples{
check_clc_year(2006)
check_clc_year(2012)
}
\keyword{internal}
