% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_buffer_composition.R
\name{extract_compositions}
\alias{extract_compositions}
\title{Extract the land cover composition}
\usage{
extract_compositions(points_df, map, buffer_radius_m, points_crs = 4326)
}
\arguments{
\item{points_df}{Data frame containing all the locations of the points from
which to extract the surrounding composition. Coordinates need to be given in
two columns names \code{longitude} and \code{latitude}. Each row must have a unique
identifier set in the \code{point_id} column.}

\item{map}{Sf object containing the polygons describing the territory land
cover.}

\item{buffer_radius_m}{Radius length in meters of the circular buffer
centered on the coordinates of the points from which to extract the land
cover composition.}

\item{points_crs}{Identifier of the coordinate reference system (CRS)
associated with the coordinates of the input points. Default is EPSG:4326.}
}
\value{
Data frame containing the relative compositions of the buffers
centered around the input points. Each land cover is associated with columns
whose values range from 0 (category absent from the buffer zone) to 1 (the
buffer is entirely composed of this category).
}
\description{
\code{extract_compositions} returns the composition of the land cover in the area
around the points given in a data frame.
}
\details{
This function performs the following steps:
\itemize{
\item Transform points: points data frame is transformed to a sf object and
converted to the CORINE data set coordinate system
\item Compute buffer compositions: for each point, a round buffer of the given
radius is intersected with the dataset land cover categories
\item Combine by point_id, land cover type and get buffer area ratio: for each
category in the buffer, the area is computed and then converted to a
proportion of the area of the buffer.
}
}
\examples{
map_folder <- system.file(
  "extdata", "2012",
  package = "corine.land.cover.landuse.extraction"
)
map <- read_clc_map_shapefile(map_folder)
buffer_radius_m <- 1
extract_compositions(
  example_points_df,
  map,
  buffer_radius_m
)
}
