% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_corine_land_cover_years.R
\name{get_clc_year}
\alias{get_clc_year}
\title{Automatically select the dataset to use}
\usage{
get_clc_year(year, available_years = getOption("available_years"))
}
\arguments{
\item{year}{Integer. Sampling year.}

\item{available_years}{Vector of available years.
Is automatically calculated with \code{\link[=set_corine_land_cover_path]{set_corine_land_cover_path()}}.}
}
\value{
Integer. Selected year.
}
\description{
\code{get_clc_year} returns the year of an available data from the given sampling
year. The selected year is the highest strictly inferior available year. If
2012 and 2018 are the available years, 2012 will be selected for date
between 2013 and 2018, 2018 will be selected for the rest. If no available
year is smaller than the sampling year, the smallest available year is
selected.
}
\examples{
get_clc_year(2002, available_years = c(2012, 2018))
get_clc_year(2013, available_years = c(2012, 2018))
get_clc_year(2022, available_years = c(2012, 2018))
}
\keyword{internal}
