% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_corine_land_cover_data.R
\name{get_data_format}
\alias{get_data_format}
\title{Get the dataset format.}
\usage{
get_data_format(data_folder)
}
\arguments{
\item{data_folder}{String. Path of the CORINE Land Cover dataset subfolder.}
}
\value{
String: 'shapefile' or 'geopackage'.
}
\description{
\code{get_data_format} gives the file format (geopackage of shapefile) of the
CORINE Land Cover dataset.
}
\examples{
folder_2012 <- system.file(
  "extdata", "2012",
  package = "corine.land.cover.landuse.extraction"
)
get_data_format(folder_2012)

folder_2018 <- system.file(
  "extdata", "2018",
  package = "corine.land.cover.landuse.extraction"
)
get_data_format(folder_2018)
}
\keyword{internal}
