% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_buffer_composition.R
\name{get_full_compositions}
\alias{get_full_compositions}
\title{Extract the land cover composition for the complete point data frame}
\usage{
get_full_compositions(
  points_df,
  buffer_radius_m,
  input_clc_year = "auto",
  points_crs = 4326
)
}
\arguments{
\item{points_df}{Data frame containing all the locations of the points from
which to extract the surrounding composition. Coordinates need to be given in
two columns names \code{longitude} and \code{latitude}. This function implies that the
coordinates are in the EPSG:4326 reference system. Each row must have a
unique identifier set in the \code{point_id} column. If the CORINE Land Cover
dataset selection is set to 'auto', a \code{year} column must be provided.}

\item{buffer_radius_m}{Radius length in meters of the circular buffer
centered on the coordinates of the points from which to extract the land
cover composition.}

\item{input_clc_year}{String 'auto' or integer. Default 'auto'. 'auto'
selects the closest previous dataset available for each point, according to
the \code{year} value. If 2012 and 2018 are the available years, 2012 will be
selected for date between 2013 and 2018, 2018 will be selected for the rest.
\code{input_clc_year} can also be an available year from which to process all data
points.}

\item{points_crs}{Identifier of the coordinate reference system (CRS)
associated with the coordinates of the input points. Default is EPSG:4326.}
}
\value{
An \code{\link[=extract_compositions]{extract_compositions()}} output. Data frame containing the
relative compositions of the buffers centered around the input points. Each
land cover is associated with columns whose values range from 0 (category
absent from the buffer zone) to 1 (the buffer is entirely composed of this
category).
}
\description{
Main function. For a given buffer radius, \code{get_full_compositions} splits
the points data according to the CORINE Land Cover dataset to use and process
to extract the composition of the land cover in the area around the points.
}
\details{
This function performs the following steps:
\itemize{
\item Compute year to use: if the year corresponding to the dataset is not
given, it is automatically calculated from the sampling year
\item Extract for each dataset: extraction is performed for each data set
\item Combine data frame: year composition data frames are combined.
}
}
\examples{
# CORINE data sets must be loaded first
clc_path <- system.file(
  "extdata", package = "corine.land.cover.landuse.extraction"
)
set_corine_land_cover_path(clc_path)
buffer_radius_m <- 1
# Extract with auto selection
get_full_compositions(
  example_points_df, buffer_radius_m, input_clc_year = "auto"
)
# Extract for a given year
get_full_compositions(
  example_points_df, buffer_radius_m, input_clc_year = 2012
)
}
