% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_buffer_composition.R
\name{get_year_compositions}
\alias{get_year_compositions}
\title{Extract the land cover composition for a given year}
\usage{
get_year_compositions(
  points_df,
  input_year,
  buffer_radius_m,
  points_crs = 4326
)
}
\arguments{
\item{points_df}{Data frame containing all the locations of the points from
which to extract the surrounding composition. Coordinates need to be given in
two columns names \code{longitude} and \code{latitude}. This function implies that the
coordinates are in the EPSG:4326 reference system. Each row must have a
unique identifier set in the \code{point_id} column. The year of the CORINE Land
Cover dataset to use must be given in the \code{clc_year} column.}

\item{input_year}{Numeric identifying the dataset to use. It is used to
filter \code{points_df}.}

\item{buffer_radius_m}{Radius length in meters of the circular buffer
centered on the coordinates of the points from which to extract the land
cover composition.}

\item{points_crs}{Identifier of the coordinate reference system (CRS)
associated with the coordinates of the input points. Default is EPSG:4326.}
}
\value{
An \code{\link[=extract_compositions]{extract_compositions()}} output. Data frame containing the
relative compositions of the buffers centered around the input points. Each
land cover is associated with columns whose values range from 0 (category
absent from the buffer zone) to 1 (the buffer is entirely composed of this
category).
}
\description{
Internal function. For a given year, \code{get_year_compositions} filters the data
points, loads the corresponding CORINE Land Cover dataset and runs the
\code{\link[=extract_compositions]{extract_compositions()}} function.
}
\details{
This function performs the following steps:
\itemize{
\item Filter: points data frame is filter to keep the locations where to extract
buffer compositions
\item Load: the corresponding CORINE Land Cover dataset is loaded inot memory
\item Extract: extraction is made between the points and the loaded dateset.
}
}
\examples{
buffer_radius_m <- 1
get_year_compositions(example_points_df, 2012, buffer_radius_m)
}
\keyword{internal}
