% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_corine_land_cover_data.R
\name{read_clc_map}
\alias{read_clc_map}
\title{Open a CORINE Land Cover dataset.}
\usage{
read_clc_map(year, clc_path = getOption("clc_path"))
}
\arguments{
\item{year}{String or integer. Year giving the subfolder to open.}

\item{clc_path}{String. Path of the folder containing all the CORINE Land
Cover dataset subfolders.}
}
\value{
Sf object.
}
\description{
\code{read_clc_map} loads a CORINE Land Cover dataset after having identified its
data type format.
}
\examples{
clc_path <- system.file(
  "extdata", package = "corine.land.cover.landuse.extraction"
)
set_corine_land_cover_path(clc_path)
read_clc_map(2012)
read_clc_map(2018)
}
\keyword{internal}
