% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_corine_land_cover_data.R
\name{read_clc_map_gpkg}
\alias{read_clc_map_gpkg}
\title{Open a geopackage dataset}
\usage{
read_clc_map_gpkg(data_folder)
}
\arguments{
\item{data_folder}{String. Path of the CORINE Land Cover dataset subfolder.}
}
\value{
Sf object.
}
\description{
\code{read_clc_map_gpkg} loads a geopackage dataset.
}
\examples{
folder_2018 <- system.file(
  "extdata", "2018",
  package = "corine.land.cover.landuse.extraction"
)
read_clc_map_shapefile(folder_2018)
}
