% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_corine_land_cover_data.R
\name{read_clc_map_shapefile}
\alias{read_clc_map_shapefile}
\title{Open a shapefile dataset}
\usage{
read_clc_map_shapefile(data_folder)
}
\arguments{
\item{data_folder}{String. Path of the CORINE Land Cover dataset subfolder.}
}
\value{
Sf object.
}
\description{
\code{read_clc_map_shapefile} loads a shapefile dataset.
}
\examples{
folder_2012 <- system.file(
  "extdata", "2012",
  package = "corine.land.cover.landuse.extraction"
)
read_clc_map_shapefile(folder_2012)
}
