% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_corine_land_cover_data.R
\name{set_corine_land_cover_path}
\alias{set_corine_land_cover_path}
\title{Initialize the dataset path}
\usage{
set_corine_land_cover_path(path)
}
\arguments{
\item{path}{String. Path to the folder containing the subfolders. Each
subfolder must contain the CORINE Land Cover dataset in geopackage or
shapefile format. The subfolder names must the years of the datasets.}
}
\description{
\code{set_corine_land_cover_path} allows the package to access the CORINE Land
Cover datasets.
}
\examples{
clc_path <- system.file(
  "extdata", package = "corine.land.cover.landuse.extraction"
)
set_corine_land_cover_path(clc_path)
}
